refactor(general_functions_pure): migrate methods
Migrated methods from general_functions_pure.py to
requests_pure.py, also modified their references.
The methods seem semantically more coherent in
requests_pure.py.
